'npm run prod' works, fully removed Storybook Some intermediate changes for Sharing Google Drive Access -- Still can't revoke sharing
Resolves https://wevoteusa.atlassian.net/issues/WV-1055
