--- .storybook/main.js
@@ -1,21 +1,0 @@
-/** @type { import('@storybook/react-webpack5').StorybookConfig } */
-const config = {
-  stories: ['../src/js/common/stories/**/*.mdx', '../src/js/common/stories/**/*.stories.@(js|jsx|ts|tsx)'],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-  ],
-  framework: {
-    name: "@storybook/react-webpack5",
-    options: {
-      builder: {
-        useSWC: true,
-      },
-    },
-  },
-  docs: {
-    autodocs: "tag",
-  },
-};
-export default config;
