+++ package.json
@@ -16,14 +16,13 @@
     "wdio": "wdio ./tests/browserstack_automation/config/wdio.config.js",
     "wdio-local-bs": "wdio ./tests/browserstack_automation/config/wdio.config.local.js",
     "wdio-local": "wdio ./tests/browserstack_automation/config/wdio.local.config.js",
-    "build": "node node/buildDateFile.js && PRODUCTION=1 MINIMIZED=1 webpack serve --mode production",
+    "build": "node node/buildDateFile.js && PRODUCTION=1 && MINIMIZED=1 webpack --mode production",
     "buildCordova": "node node/buildDateFile.js && node node/buildSrcCordova && CORDOVA=1 && webpack --mode development && node node/logCompileDate.js",
     "buildCordovaAndLinks": "node node/buildDateFile.js && node node/buildSrcCordova && CORDOVA=1 && webpack --mode development && node node/buildSymLinksRemote.js && bash ./node/unSymLinkIOS.sh && node node/logCompileDate.js",
     "lint": "eslint --format stylish --ext .jsx --ext .js src/js",
     "lintCordova": "eslint --format stylish --ext .jsx --ext .js srcCordova/js",
     "prod": "npm run build",
     "run-prod-local": "node server-prod-local.js",
-    "start-with-storybook": "node node/buildDateFile.js && npm run build-storybook && PROTOCOL=HTTP webpack serve --mode development",
     "start": "node node/buildDateFile.js && PROTOCOL=HTTP webpack serve --mode development",
     "start-win": "node node/buildDateFile.js && webpack serve --mode development",
     "start-https": "node node/buildDateFile.js && PROTOCOL=HTTPS webpack serve --mode development",
@@ -37,8 +36,6 @@
     "autoTest": " mocha --require @babel/register --require tests/component/index.js \"./tests/component/tests.js\" && npm run lint",
     "snyk-protect": "snyk-protect",
     "prepare": "npm run snyk-protect",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build",
     "chromatic": "npx chromatic --project-token=chpt_89afc2af23dd1fd"
   },
   "devDependencies": {
@@ -67,7 +64,6 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-storybook": "^0.6.15",
     "eslint-webpack-plugin": "^2.4.3",
     "express": "^5.0.0-beta.3",
     "file-loader": "^6.2.0",
@@ -90,9 +86,9 @@
     "@dnd-kit/modifiers": "^9.0.0",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@mui/icons-material": "~5.17.1",
+    "@mui/icons-material": "^5.17.1",
     "@mui/lab": "^5.0.0-alpha.96",
-    "@mui/material": "~5.17.1",
+    "@mui/material": "^5.16.0",
     "@mui/styled-engine-sc": "^6.3.0",
     "@mui/styles": "^5.16.0",
     "@mui/x-date-pickers": "^7.27.0",
@@ -115,6 +111,7 @@
     "keymirror": "~0.1.1",
     "lodash-es": "^4.17.21",
     "moment": "^2.29.4",
+    "mui-file-dropzone": "^4.0.1",
     "object-assign": "~4.1.1",
     "process": "^0.11.10",
     "prop-types": "~15.8.1",
@@ -138,11 +135,6 @@
     "topojson-client": "^3.1.0",
     "use-file-picker": "^2.1.2",
     "validator": "^13.12.0"
-  },
-  "eslintConfig": {
-    "extends": [
-      "plugin:storybook/recommended"
-    ]
   },
   "bugs": {
     "url": "https://github.com/wevote/weconnect/issues"
